+++ src/domain/communication/discussion/components/CategorySelector.tsx
